apps/voice/pyproject.toml
[project]
name = "engram-voice"
version = "0.1.0"
description = "Voice processing service for ENGRAM - Pipecat with Silero VAD, faster-whisper STT, and Chatterbox TTS"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pipecat-ai>=0.0.30",
    "faster-whisper>=1.0.0",
    "silero-vad>=5.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]
chatterbox = [
    "chatterbox-tts>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
strict = true
